Convert to a single instance for library use
