+++ Cargo.toml
@@ -9,9 +9,4 @@ edition = "2024"
 [workspace.dependencies]
 zeevonk = { path = "crates/zeevonk" }
 
-theymx = "0.1.1"
 log = "0.4.29"
-thiserror = "2.0"
-serde = "1.0"
-uuid = "1.19.0"
-tokio = "1.49.0"
