+++ crates/cli/Cargo.toml
@@ -4,12 +4,10 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
-zeevonk = { workspace = true, features = ["server"] }
+zeevonk.workspace = true
 
 log.workspace = true
-
-tokio = { workspace = true, features = ["rt", "rt-multi-thread"] }
-
 pretty_env_logger = "0.5.0"
+
 anyhow = "1.0"
 clap = { version = "4.5.53", features = ["derive"] }
