+++ crates/cli/src/run.rs
@@ -1,14 +1,14 @@
 use std::path::PathBuf;
 
-use zeevonk::{project::file::ProjectFile, server::Server};
+use zeevonk::{Zeevonk, project::file::ProjectFile};
 
 /// Runs the project at the given path.
-pub async fn run_project(project_path: PathBuf) -> anyhow::Result<()> {
+pub fn run_project(project_path: PathBuf) -> anyhow::Result<()> {
     let project_definition = ProjectFile::load_from_folder(&project_path)?;
-    let server = Server::new(project_definition)?;
-    server.start().await?;
+    let zeevonk = Zeevonk::new(project_definition)?;
+    zeevonk.start();
 
     loop {
-        std::thread::sleep(std::time::Duration::from_secs_f32(1.0 / 60.0));
+        std::thread::sleep(std::time::Duration::from_secs_f32(1.0 / 30.0));
     }
 }
