+++ crates/zeevonk/Cargo.toml
@@ -11,45 +11,17 @@ categories = ["multimedia", "network-programming"]
 version.workspace = true
 edition.workspace = true
 
-[features]
-server = [
-    "dep:tokio",
-    "dep:tokio-tungstenite",
-    "dep:futures-util",
-    "dep:spin_sleep",
-    "dep:crossbeam-channel",
-    "dep:libftd2xx",
-    "dep:gdtf",
-    "dep:arrayvec",
-    "dep:socket2",
-]
-client = [
-    "dep:tokio",
-    "dep:tokio-tungstenite",
-    "dep:futures-util",
-]
-
 [dependencies]
 log.workspace = true
-theymx = { workspace = true, features = ["serde"] }
-thiserror.workspace = true
-serde = { workspace = true, features = ["derive"] }
-uuid = { workspace = true, features = ["serde",  "v4"] }
-
+theymx = { version = "0.1.1", features = ["serde"] }
+thiserror = "2.0"
+serde = { version = "1.0", features = ["derive"] }
+uuid = { version = "1.19.0", features = ["serde", "v4"] }
 lazy_static = "1.5.0"
 serde_json = "1.0.149"
-tokio = { workspace = true, optional = true, features = ["sync", "rt", "macros"] }
-tokio-tungstenite = { version = "0.28.0", optional = true }
-
-# feature = "server"
-spin_sleep = { version = "1.3.3", optional = true }
-crossbeam-channel = { version = "0.5.15", optional = true }
-libftd2xx = { version = "0.33.1", optional = true }
-gdtf = { version = "0.2.0", optional = true }
-tungstenite = { version = "0.28.0", optional = true }
-futures-util = { version = "0.3.31", optional = true }
-arrayvec = { version = "0.7.6", optional = true }
-socket2 = { version = "0.6.1", optional = true }
-
-[dev-dependencies]
-pretty_env_logger = "0.5.0"
+spin_sleep = "1.3.3"
+crossbeam-channel = "0.5.15"
+libftd2xx = "0.33.1"
+gdtf = "0.2.0"
+arrayvec = "0.7.6"
+socket2 = "0.6.1"
