+++ crates/zeevonk/src/lib.rs
@@ -1,36 +1,92 @@
 #![warn(missing_docs)]
 
-//! Zeevonk is a modular lighting control system.
-//! The Zeevonk server is a hub for multiple clients that can update attribute values,
-//! which the server then will output to the configured DMX protocols.
-//!
-//! <div class="warning">
-//!
-//! **Warning**
-//!
-//! Zeevonk is currently in early development. APIs, features, and behavior may change frequently and without notice.
-//! It is not yet recommended for production use.
-//!
-//! </div>
-//!
-//! # Components
-//!
-//! - [Server](crate::server): Manages clients, resolves attributes, and sends DMX.
-//! - [Client](crate::client): Can generate and send attribute values to the server, or talk to other clients using [`Trigger`][trigger::Trigger]s.
-//!
-//! See the respective module documentation for details.
+//! Zeevonk.
 
 pub mod attr;
 pub mod error;
-pub mod ident;
-pub mod packet;
 pub mod project;
-pub mod trigger;
 pub mod value;
 
-#[cfg(any(feature = "client"))]
-pub mod client;
-#[cfg(feature = "server")]
-pub mod server;
+mod output;
+mod project_builder;
+mod resolver;
+
+use std::sync::Arc;
+
+use crate::attr::Attribute;
+use crate::output::agent::OutputAgent;
+use crate::project::Project;
+use crate::project::file::ProjectFile;
+use crate::project::stage::FixtureId;
+use crate::value::{AttributeValues, ClampedValue};
 
 pub use error::{Error, Result};
+
+/// The main entry point for interacting with a Zeevonk instance.
+pub struct Zeevonk {
+    project: Arc<Project>,
+    output_agent: Arc<OutputAgent>,
+}
+
+impl Zeevonk {
+    /// Creates a new [`Zeevonk`] instance from a [`ProjectFile`].
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if the project file cannot be loaded or parsed.
+    pub fn new(project_file: ProjectFile) -> crate::Result<Self> {
+        let project_handle = Arc::new(project_builder::from_file(project_file)?);
+        let output_agent = Arc::new(OutputAgent::new(project_handle.clone()));
+        Ok(Self { project: Arc::clone(&project_handle), output_agent })
+    }
+
+    /// Starts Zeevonk.
+    pub fn start(&self) {
+        self.output_agent.start();
+        log::info!("Zeevonk started");
+    }
+
+    /// Returns a reference to the loaded [`Project`].
+    pub fn project(&self) -> &Project {
+        &self.project
+    }
+
+    /// Sets attribute values for fixtures in the project.
+    ///
+    /// This method updates the output agent with the provided values. If `include_children` is
+    /// `true`, the values are also applied recursively to all child fixtures.
+    pub fn set_attribute_values(&self, values: AttributeValues, include_children: bool) {
+        log::debug!("setting attribute values | include_children={}", include_children);
+        self.output_agent.update_values(values.clone());
+
+        if include_children {
+            /// Recursively sets attribute values for child fixtures.
+            fn set_values_recursively(
+                project: &Project,
+                output_agent: &OutputAgent,
+                fixture_id: &FixtureId,
+                attribute: Attribute,
+                value: ClampedValue,
+            ) {
+                let Some(fixture) = project.stage().fixtures().get(fixture_id) else {
+                    return;
+                };
+
+                for sub_id in fixture.sub_ids() {
+                    output_agent.update_value(*sub_id, attribute, value);
+                    set_values_recursively(project, output_agent, &sub_id, attribute, value);
+                }
+            }
+
+            for (fixture_id, attribute, value) in values.values() {
+                set_values_recursively(
+                    &self.project,
+                    &self.output_agent,
+                    fixture_id,
+                    *attribute,
+                    *value,
+                );
+            }
+        }
+    }
+}
