+++ crates/zeevonk/src/output/protocols/sacn/mod.rs
@@ -6,7 +6,7 @@ use arrayvec::ArrayVec;
 use theymx::{Multiverse, UniverseId};
 use uuid::Uuid;
 
-use crate::server::output;
+use crate::output;
 
 mod acn;
 mod packet;
